refining inventory and api controllers: created v1 folder, add a new InventoryController, moved 3 endpoints from BFFApiGatewayController,and 1 from v2 as well
