+++ api-gateway/src/main/java/com/petclinic/bffapigateway/presentationlayer/v2/InventoryController.java
@@ -222,15 +222,7 @@ public class InventoryController {
                 .defaultIfEmpty(ResponseEntity.badRequest().build());
     }
 
-    @SecuredEndpoint(allowedRoles = {Roles.ADMIN, Roles.INVENTORY_MANAGER})
-    @PostMapping("/{inventoryId}/products")
-    public Mono<ResponseEntity<ProductResponseDTO>> addSupplyToInventory(
-            @RequestBody ProductRequestDTO productRequestDTO,
-            @PathVariable String inventoryId) {
-        return inventoryServiceClient.addSupplyToInventory(productRequestDTO, inventoryId)
-                .map(productResponseDTO -> ResponseEntity.status(HttpStatus.CREATED).body(productResponseDTO))
-                .defaultIfEmpty(ResponseEntity.badRequest().build());
-    }
+
 
     @SecuredEndpoint(allowedRoles = {Roles.ADMIN, Roles.INVENTORY_MANAGER})
     @GetMapping("/{inventoryId}/productquantity")
