+++ petclinic-frontend/src/features/inventories/api/addInventory.ts
@@ -6,7 +6,9 @@ export default async function addInventory(
 ): Promise<void> {
   try {
     // Append the appropriate endpoint for adding an inventory
-    await axiosInstance.post<void>('/inventories', inventoryData);
+    await axiosInstance.post<void>('/inventory', inventoryData, {
+      useV2: false,
+    });
   } catch (error) {
     console.error('Error adding inventory:', error);
     throw error;
