+++ petclinic-frontend/src/features/inventories/api/getAllInventoryTypes.ts
@@ -3,7 +3,10 @@ import { InventoryType } from '@/features/inventories/models/InventoryType.ts';
 
 export async function getAllInventoryTypes(): Promise<InventoryType[]> {
   const response = await axiosInstance.get<InventoryType[]>(
-    axiosInstance.defaults.baseURL + 'v2/gateway/inventories/types'
+    '/inventory/types',
+    {
+      useV2: false,
+    }
   );
   return response.data;
 }
